publish-schema/src/requirements.txt
cloudevents==1.11.0
google-auth
google-cloud-secret-manager
google-cloud-pubsub
requests
pydantic_settings==2.1.0
python-dotenv==1.0.0
httpx==0.23.3
jsonschema==4.17.3
functions-framework==3.5.0
pip_audit==2.5.6
setuptools==70.0.0